Make rand seq function exported by chunk package (this probably should get refactored at some point into a test utility package)
