+++ src/server/pkg/storage/chunk/util.go
@@ -1,6 +1,7 @@
 package chunk
 
 import (
+	"math/rand"
 	"os"
 	"testing"
 
@@ -21,4 +22,15 @@ func LocalStorage(tb testing.TB) (obj.Client, *Storage) {
 	objC, err := obj.NewLocalClient(wd)
 	require.NoError(tb, err)
 	return objC, NewStorage(objC, Prefix)
+}
+
+var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
+
+// RandSeq generates a random sequence of data (n is number of bytes)
+func RandSeq(n int) []byte {
+	b := make([]rune, n)
+	for i := range b {
+		b[i] = letters[rand.Intn(len(letters))]
+	}
+	return []byte(string(b))
 }
